Move NodeUnit to framework, and update include paths. Fix order of nodes in partition. DQ -> target -> Q must be added sequentially. Fix x86 build error. Fix android issue with gtest flags not being processed (couldn't debug failing test easily when --gtest_filter didn't work)
