+++ onnxruntime/core/providers/xnnpack/detail/utils.cc
@@ -6,14 +6,14 @@
 #include <vector>
 
 #include "core/common/common.h"
+#include "core/common/safeint.h"
+#include "core/framework/node_unit.h"
 #include "core/framework/tensorprotoutils.h"
 #include "core/graph/indexed_sub_graph.h"
 #include "core/graph/node_attr_utils.h"
-
-#include "core/providers/shared/node_unit/node_unit.h"
-#include "onnx/defs/attr_proto_util.h"
-#include "core/common/safeint.h"
 #include "core/optimizer/initializer.h"
+
+#include "onnx/defs/attr_proto_util.h"
 
 namespace onnxruntime {
 namespace xnnpack {
